+++ frontend/src/components/record/MemberRecordList.tsx
@@ -1,10 +1,8 @@
-import { styled } from 'styled-components';
-
 import Accordion from '@Components/common/Accordion/Accordion';
+import AccordionSkeleton from '@Components/common/Accordion/AccordionSkeleton';
 import Typography from '@Components/common/Typography/Typography';
 
 import MemberRecord from './MemberRecord';
-import { TextSkeletonStyle } from './StudyInformation';
 
 type Props = {
   members?: { memberId: number; nickname: string }[];
@@ -34,24 +32,3 @@ const MemberRecordList = ({ members = [], studyId = '', isLoading }: Props) => {
 };
 
 export default MemberRecordList;
-
-const AccordionSkeleton = () => {
-  return (
-    <AccordionSkeletonLayout>
-      <SkeletonItem />
-      <SkeletonItem />
-    </AccordionSkeletonLayout>
-  );
-};
-
-const AccordionSkeletonLayout = styled.div`
-  display: grid;
-  row-gap: 40px;
-
-  max-width: 1200px;
-`;
-
-const SkeletonItem = styled.div`
-  height: 80px;
-  ${TextSkeletonStyle}
-`;
